Reorganized into the directory structure I want
