--- src/NUnitEngine/Portable/nunit.portable.agent/Internal/ILogger.cs
@@ -1,91 +1,0 @@
-﻿// ***********************************************************************
-// Copyright (c) 2012 Charlie Poole
-//
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to
-// permit persons to whom the Software is furnished to do so, subject to
-// the following conditions:
-// 
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-// 
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// ***********************************************************************
-
-using System;
-
-#if NUNIT_ENGINE
-namespace NUnit.Engine
-#elif NUNIT_FRAMEWORK
-namespace NUnit.Framework.Internal
-#else
-namespace NUnit.Common
-#endif
-{
-    /// <summary>
-    /// Interface for logging within the engine
-    /// </summary>
-    public interface ILogger
-    {
-        /// <summary>
-        /// Logs the specified message at the error level.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        void Error(string message);
-
-        /// <summary>
-        /// Logs the specified message at the error level.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        /// <param name="args">The arguments.</param>
-        void Error(string message, params object[] args);
-
-        /// <summary>
-        /// Logs the specified message at the warning level.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        void Warning(string message);
-
-        /// <summary>
-        /// Logs the specified message at the warning level.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        /// <param name="args">The arguments.</param>
-        void Warning(string message, params object[] args);
-
-        /// <summary>
-        /// Logs the specified message at the info level.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        void Info(string message);
-
-        /// <summary>
-        /// Logs the specified message at the info level.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        /// <param name="args">The arguments.</param>
-        void Info(string message, params object[] args);
-
-        /// <summary>
-        /// Logs the specified message at the debug level.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        void Debug(string message);
-
-        /// <summary>
-        /// Logs the specified message at the debug level.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        /// <param name="args">The arguments.</param>
-        void Debug(string message, params object[] args);
-    }
-}
