+++ src/nunit.portable.agent/Common/ColorConsoleWriter.cs
