+++ src/nunit.portable.agent/Common/ColorStyle.cs
