+++ src/nunit.portable.agent/Common/OutputSpecification.cs
