+++ src/nunit.portable.agent/Common/PackageSettings.cs
