+++ src/nunit.portable.agent/Common/TestSelectionParser.cs
@@ -28,11 +28,7 @@ using System.Text;
 // Missing XML Docs
 #pragma warning disable 1591
 
-#if NUNIT_ENGINE
 namespace NUnit.Engine
-#else
-namespace NUnit.Common
-#endif
 {
     public class TestSelectionParser
     {
