+++ src/nunit.portable.agent/Common/Tokenizer.cs
@@ -27,11 +27,7 @@ using System.Text;
 // Missing XML Docs
 #pragma warning disable 1591
 
-#if NUNIT_ENGINE
 namespace NUnit.Engine
-#else
-namespace NUnit.Common
-#endif
 {
     public enum TokenKind
     {
