+++ src/nunit.portable.agent/Compatibility/Path.cs
