+++ src/nunit.portable.agent/Properties/AssemblyInfo.cs
