+++ src/nunit.portable.agent/ResultSummary.cs
