+++ src/nunit.portable.agent/nunit.snk
