+++ test/mock-assembly/Properties/AssemblyInfo.cs
