+++ test/mock-assembly/packages.config
