+++ test/nunit.portable.agent.tests/Properties/AssemblyInfo.cs
