+++ test/nunit.portable.agent.tests/ResultSummaryTests.cs
