+++ test/nunit.portable.agent.tests/XmlHelper.cs
