+++ test/nunit.portable.agent.tests/nunit.portable.agent.tests.csproj
@@ -17,8 +17,8 @@
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
-    <OutputPath>..\..\..\..\bin\Debug\agents\</OutputPath>
-    <DefineConstants>TRACE;DEBUG;NUNIT_ENGINE;PORTABLE</DefineConstants>
+    <OutputPath>bin\Debug\</OutputPath>
+    <DefineConstants>TRACE;DEBUG;PORTABLE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <Prefer32Bit>false</Prefer32Bit>
@@ -26,15 +26,15 @@
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
     <Optimize>true</Optimize>
-    <OutputPath>..\..\..\..\bin\Release\agents\</OutputPath>
-    <DefineConstants>TRACE;NUNIT_ENGINE;PORTABLE</DefineConstants>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE;PORTABLE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="nunit.framework, Version=3.2.1.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
-      <HintPath>..\..\..\..\packages\NUnit.3.2.1\lib\net45\nunit.framework.dll</HintPath>
+      <HintPath>..\..\packages\NUnit.3.2.1\lib\net45\nunit.framework.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="System" />
@@ -43,7 +43,7 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\..\..\CommonAssemblyInfo.cs">
+    <Compile Include="..\..\src\CommonAssemblyInfo.cs">
       <Link>Properties\CommonAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="NUnitPortableDriverTests.cs" />
@@ -52,22 +52,22 @@
     <Compile Include="XmlHelper.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\..\NUnitFramework\mock-assembly\mock-assembly-4.5.csproj">
-      <Project>{c3264ff3-e2d7-48d7-93a7-cdc9a827ad9d}</Project>
-      <Name>mock-assembly-4.5</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\nunit.portable.agent\nunit.portable.agent.csproj">
-      <Project>{0f18f7a5-87ca-41e0-acb2-a13c68d6cc5b}</Project>
-      <Name>nunit.portable.agent</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="MockTestResult.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\nunit.portable.agent\nunit.portable.agent.csproj">
+      <Project>{0f18f7a5-87ca-41e0-acb2-a13c68d6cc5b}</Project>
+      <Name>nunit.portable.agent</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\mock-assembly\mock-assembly-4.5.csproj">
+      <Project>{c3264ff3-e2d7-48d7-93a7-cdc9a827ad9d}</Project>
+      <Name>mock-assembly-4.5</Name>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
