+++ test/nunit.portable.agent.tests/packages.config
